Refactoring of intake monitor into inner class in intake
